refactor(#1682): rewrite refs reyn.search_backends → reyn.tools.search_backends (A3a)
2 files, zero bare remain. class-3 clean; no __file__ walks. ruff clean.
